+++ cardigan/build-cardigan.js
@@ -1,7 +1,7 @@
 require('babel-polyfill');
 require('babel-register');
 
-const fractal = require('./fractal-app').default;
+const fractal = require('./fractal');
 const builder = fractal.web.builder();
 
 console.info('Buttoning up Cardigan...');
