+++ cardigan/cardigan-theme/assets/favicon.ico
