+++ cardigan/cardigan-theme/partials/browser/panel-notes.nunj
