+++ cardigan/docs/BodyBlock.config.js
