+++ cardigan/docs/visual-architecture.jsx
