+++ cardigan/package.json
@@ -1,21 +1,21 @@
 {
+  "name": "cardigan",
+  "version": "1.0.0",
   "description": "Wellcome Collection's design system",
-  "repository": "https://github.com/wellcometrust/wellcomecollection.org/tree/master/cardigan",
-  "license": "MIT",
+  "main": "index.js",
   "scripts": {
-    "app": "fractal start --sync --port 3002",
-    "app:build": "node ./build-cardigan.js",
-    "app:dev": "nodemon -e js,njk,yml,md --exec 'fractal start --sync --port 3002'"
+    "dev": "fractal start",
+    "build": "node ./build-cardigan.js"
   },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
   "dependencies": {
     "@frctl/fractal": "^1.1.7",
-    "@frctl/mandelbrot": "^1.2.0",
-    "@frctl/nunjucks": "^1.0.3",
-    "babel-polyfill": "^6.26.0",
+    "babel-plugin-transform-react-constant-elements": "^6.23.0",
+    "babel-plugin-transform-react-inline-elements": "^6.22.0",
     "babel-preset-env": "^1.6.1",
-    "babel-register": "^6.26.0",
-    "fractal": "0.0.3",
-    "immutable": "^3.8.2",
-    "nodemon": "^1.12.1"
+    "babel-preset-react": "^6.24.1",
+    "fractal-react-adapter": "https://github.com/davidpmccormick/fractal-react-adapter.git"
   }
 }
