--- cardy/.babelrc
@@ -1,10 +1,0 @@
-{
-  "presets": [
-    "env",
-    "react"
-  ],
-  "plugins": [
-    "transform-react-constant-elements",
-    "transform-react-inline-elements"
-  ]
-}
