--- cardy/config/statuses.js
@@ -1,27 +1,0 @@
-module.exports = {
-  wip: {
-    label: 'WIP',
-    description: 'Work in progress',
-    color: '#ad4e00'
-  },
-  testing: {
-    label: 'Testing',
-    description: 'Being tested',
-    color: '#895791'
-  },
-  graduated: {
-    label: 'Graduated',
-    description: 'Implemented',
-    color: '#367378'
-  },
-  deprecated: {
-    label: 'Deprecated',
-    description: 'Deprecated',
-    color: '#c72e3d'
-  },
-  benched: {
-    label: 'Benched',
-    description: 'Has previously been used, but is not currently; maybe be used in future',
-    color: '#006272'
-  }
-};
